feat(config): gif recording length configuration value (#98)
Closes #74

Introduce a way to give the developer the ability to extend the length of the gif capture feature.

I sadly kinda saw the environment variable propose to the solution late, but I don't fully agree wit it either. The ability to extend gif length to showcase your work/let people send you cool moments over 5 seconds I feel like should be an accessible option just as any other configuration.

Reviewed-on: https://codeberg.org/brettchalupa/usagi/pulls/98

+++ src/font_bake.rs
@@ -260,7 +260,7 @@ pub fn run(
         // Unpack FT_PIXEL_MODE_MONO. Each row is `pitch` bytes,
         // packed MSB-first (bit 7 = leftmost pixel). Pitch can include
         // padding bytes past width/8 rounded up.
-        let pitch = bitmap.pitch() as i32;
+        let pitch = bitmap.pitch();
         let buf = bitmap.buffer();
         let abs_pitch = pitch.unsigned_abs() as usize;
         let mut pixels: Vec<bool> = Vec::with_capacity((w * h) as usize);
